Move controllers from main script to experiment.
