+++ flask-server/auth/__init__.py
@@ -4,4 +4,3 @@ auth = Blueprint('auth', __name__)
 
 
 from . import views
-
