--- flask-server/settings.py
@@ -1,48 +1,0 @@
-#! python3
-# -*- encoding: utf-8 -*-
-'''
-@File    :   settings.py
-@Created :   2023/10/29 21:44
-@Author  :   Zhong, Yinjie
-@Version :   1.0
-@Contact :   yinjie.zhong@vanderbilt.edu
-'''
-
-# Here put the import lib.
-
-
-# Configuration file.
-import os
-
-__basedir=os.getcwd()
-
-ENV='development'
-DEBUG=True
-SECRET_KEY='9131-0120-MA1H' # A custom value but mandatory.
-
-# Enable Non-ASCII Characters.
-JSON_AS_ASCII =False
-JSONIFY_MIMETYPE = "application/json;charset=utf-8"
-
-# Database.
-SQLALCHEMY_DATABASE_URI=f'sqlite:///{os.path.join(__basedir, "instance", "enzyhtp_gpt.db")}'
-SQLALCHEMY_TRACK_MODIFICATIONS=False
-
-# Uri
-OAUTH_VENDOR_LOGIN_CALLBACK_REDIRECT_URI = "/api/auth/profile"
-OPENAI_API_URI = "https://api.openai.com/v1/chat/completions"
-
-# Email for password reset.
-MAIL_SERVER = "smtp.gmail.com"
-MAIL_PORT = 465
-MAIL_USE_SSL = True
-MAIL_USE_TLS = False
-MAIL_USERNAME = "website.enzyhtp@gmail.com"
-MAIL_DEFAULT_SENDER = ("EnzyHTP Web Application", "website.enzyhtp@gmail.com")
-MAIL_PASSWORD = "ymyiwgzhxxpnlqcg"
-# https://mailtrap.io/blog/python-send-email-gmail/
-
-MAIL_PASSWORD_RESET_HTML_TEMPLATE = open(os.path.join(__basedir, "templates", "password_reset_email.html")).read()
-
-APP_HOST = "localhost"
-
